requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==4.0.2
django-bootstrap5==21.3
django-environ==0.8.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
python-decouple==3.6
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
